% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeValidCode.R
\name{checkInValid}
\alias{checkInValid}
\title{Check the validity of code}
\usage{
checkInValid(temp)
}
\arguments{
\item{temp}{A character string to check validity}
}
\description{
Check the validity of code
}
