% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrAssist.R
\name{dplyrAssist}
\alias{dplyrAssist}
\title{A shiny app for learn dplyr}
\usage{
dplyrAssist(df = NULL, right = NULL)
}
\arguments{
\item{df}{A tibble or a tbl_df or a data.frame to manipulate}

\item{right}{Optional second data(A tibble or a tbl_df or a data.frame) to join}
}
\value{
A manipulated tibble or NULL
}
\description{
A shiny app for learn dplyr
}
\examples{
library(tidyverse)
library(nycflights13)
#dplyrAssist(band_members,band_instruments)
#dplyrAssist(flights)
#result<-dplyrAssist(iris)
#cat(attr(result,"code"))
}
