% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeValidCode.R
\name{findData}
\alias{findData}
\title{Detect the valid data}
\usage{
findData(codes)
}
\arguments{
\item{codes}{A character vector to detect}
}
\value{
The last data in codes
}
\description{
Detect the valid data
}
\examples{
require(tidyverse)
temp <- "iris \%>\%
group_by(Species) \%>\%
summarize_all(mean)

table1"
findData(temp)
}
