% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeValidCode.R
\name{makeValid}
\alias{makeValid}
\title{Make valid R code}
\usage{
makeValid(temp)
}
\arguments{
\item{temp}{A character vector to make a valid code}
}
\value{
A valid R codes
}
\description{
Make valid R code
}
\examples{
require(tidyverse)
temp <- "iris \%>\%
group_by(Species) \%>\%
summarize_all(mean)

table1"
makeValid(temp)
}
