% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydplyr.R
\name{mydplyr}
\alias{mydplyr}
\title{A shiny app for learn dplyr}
\usage{
mydplyr(df, right = NULL)
}
\arguments{
\item{df}{A tibble or a tbl_df or a data.frame to manipulate}

\item{right}{Optional second data(A tibble or a tbl_df or a data.frame) to join}
}
\value{
A manipulated tibble or NULL
}
\description{
A shiny app for learn dplyr
}
\examples{
library(tidyverse)
library(nycflights13)
#mydplyr(band_members,band_instruments)
#mydplyr(flights)
#result<-mydplyr(iris)
#cat(attr(result,"code"))
}
